requirements.txt
Package                Version
---------------------- --------
amqp                   5.2.0
asgiref                3.6.0
async-timeout          4.0.3
billiard               4.2.0
bleach                 6.0.0
celery                 5.3.6
click                  8.1.7
click-didyoumean       0.3.0
click-plugins          1.1.1
click-repl             0.3.0
colorama               0.4.6
defusedxml             0.7.1
diff-match-patch       20230430
Django                 4.1
django-admin-interface 0.26.1
django-colorfield      0.10.1
django-cors-headers    4.1.0
django-environ         0.11.2
django-froala-editor   4.1.4
django-import-export   3.3.7
django-mathfilters     1.0.0
django-summernote      0.8.20.0
et-xmlfile             1.1.0
kombu                  5.3.5
MarkupPy               1.14
mysqlclient            2.2.0
odfpy                  1.4.1
openpyxl               3.1.2
Pillow                 9.5.0
pip                    23.3.2
prompt-toolkit         3.0.43
python-dateutil        2.8.2
python-slugify         8.0.1
PyYAML                 6.0.1
redis                  5.0.1
setuptools             69.0.3
six                    1.16.0
sqlparse               0.4.4
tablib                 3.5.0
text-unidecode         1.3
typing_extensions      4.9.0
tzdata                 2024.1
vine                   5.1.0
wcwidth                0.2.13
webencodings           0.5.1
wheel                  0.42.0
whitenoise             6.5.0
xlrd                   2.0.1
xlwt                   1.3.0
